% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_y_axis.R
\name{blank_y_axis}
\alias{blank_y_axis}
\title{Blank Y-axis}
\usage{
blank_y_axis(...)
}
\arguments{
\item{...}{Additional arguments to pass to `theme()`. Defaults is none.}
}
\description{
Remove the Y axis and all markings from the plot
}
