% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x_axis.R
\name{rotate_x_axis}
\alias{rotate_x_axis}
\title{Rotate x-axis text}
\usage{
rotate_x_axis(...)
}
\arguments{
\item{...}{Arguments to pass to `element_text()`. Default is `angle = 90, hjust = 1`.}
}
\description{
Utility to easily rotate x-axis text
}
