% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_y_axis.R
\name{rotate_y_axis}
\alias{rotate_y_axis}
\title{Rotate y-axis text}
\usage{
rotate_y_axis(...)
}
\arguments{
\item{...}{Arguments to pass to `element_text()`. Default is `angle = 90, vjust = 0`.}
}
\description{
Utility to easily rotate y-axis text
}
