% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_months.R
\name{scale_x_months}
\alias{scale_x_months}
\title{Set x-axis labels to month names}
\usage{
scale_x_months(...)
}
\arguments{
\item{...}{Additional arguments to pass to `scale_x_continuous()`. Default is none.}
}
\description{
Utility to easily replace x-axis 1:12 labels with month names
}
