% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_percent.R
\name{scale_x_percent}
\alias{scale_x_percent}
\title{Show x-axis on percent scale}
\usage{
scale_x_percent(...)
}
\arguments{
\item{...}{Additional arguments to pass to `scale_x_continuous()`. Default is none.}
}
\description{
Utility to easily convert x-axis to percentage scale
}
