% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_months.R
\name{scale_y_months}
\alias{scale_y_months}
\title{Set y-axis labels to month names}
\usage{
scale_y_months(...)
}
\arguments{
\item{...}{Additional arguments to pass to `scale_y_continuous()`. Default is none.}
}
\description{
Utility to easily replace y-axis 1:12 labels with month names
}
