% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_percent.R
\name{scale_y_percent}
\alias{scale_y_percent}
\title{Show y-axis on percent scale}
\usage{
scale_y_percent(...)
}
\arguments{
\item{...}{Additional arguments to pass to `scale_y_continuous()`. Default is none.}
}
\description{
Utility to easily convert y-axis to percentage scale
}
