% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x_axis.R
\name{rotate_x_axis}
\alias{rotate_x_axis}
\title{Rotate x-axis text}
\usage{
rotate_x_axis(angle = 90, hjust = 1)
}
\arguments{
\item{angle}{Angle to rotate x-axis text}

\item{hjust}{Horizontal justification}
}
\description{
Utility to easily rotate x-axis text
}
